Sun, Jul 2, 2023, 7:58 PM +03:00

+++ sentimental-cash/cash.py
@@ -1,5 +1,5 @@
 def main():
-    cents = int(input())
+    cents = float(input())
 
 
     quarters = calculate_quarters(cents)
